formating and int error in PG

+++ pypulse/PulseGeneration.py
@@ -1,6 +1,5 @@
 import scipy.signal as signal
 import numpy as np
-import matplotlib.pyplot as plt
 import scipy.io as sio
 
 
@@ -329,8 +328,8 @@ def dummy_noise_pulse(sampling_rate, params):
     pulse = (np.array(signal.square(2 * np.pi * params['shatter_frequency'] * t, duty=guide_pulse)) / 2) + 0.5
 
     # Attach onset and offset
-    onset = np.zeros(sampling_rate * params['onset'])
-    offset = np.zeros(sampling_rate * params['offset'])
+    onset = np.zeros(int(sampling_rate * params['onset']))
+    offset = np.zeros(int(sampling_rate * params['offset']))
 
     total_length = round(duration + params['onset'] + params['offset'], 10)
     return np.hstack((onset, pulse, offset)), np.linspace(0, total_length, total_length * sampling_rate)
@@ -354,22 +353,3 @@ def multi_noise_pulse(sampling_rate, global_onset, global_offset, params_list):
     t = np.linspace(0, pulse_matrix.shape[1] / sampling_rate, pulse_matrix.shape[1])
 
     return pulse_matrix, t
-
-
-# params = {'fromLength': True, 'fromRepeats': False, 'frequency': 20, 'repeats': 5, 'seed': 1, 'amp_min': 0.1,
-#           'amp_max': 0.9, 'shatter_frequency': 500, 'length': 1, 'onset': 0.1, 'offset': 0.1}
-#
-# pulse, t = dummy_noise_pulse(10000, params)
-
-# plt.plot(t, pulse)
-# plt.show()
-
-# Testing - random shatter
-# pulse, t = random_shatter_pulse(20000.0, 2.0, 5.0, 0.5, 200.0, 1.0, 0.1, 0.9)
-# print(sum(pulse) / len(pulse))
-# plt.plot(t, pulse)
-# plt.xlim((-0.1, 2.1))
-# plt.ylim((-0.1, 1.1))
-# plt.show()
-
-
